+++ src/resources/views/partials/lang-selector-form.blade.php
@@ -7,17 +7,17 @@
     <div class="tab-pane active" id="lang-en{{ isset($selector_name) ? '-' . $selector_name : '' }}">
         <div class="form-group">
             {!! Form::label('name', Lang::get('redminportal::forms.title')) !!}
-            {!! Form::text('name', null, array('class' => 'form-control')) !!}
+            {!! Form::text('name', (isset($translatable) ? $translatable->name : null), array('class' => 'form-control')) !!}
         </div>
 
         <div class="form-group">
             {!! Form::label('short_description', Lang::get('redminportal::forms.summary')) !!}
-            {!! Form::text('short_description', null, array('class' => 'form-control')) !!}
+            {!! Form::text('short_description', (isset($translatable) ? $translatable->short_description : null), array('class' => 'form-control')) !!}
         </div>
 
         <div class="form-group">
             {!! Form::label('long_description', Lang::get('redminportal::forms.description')) !!}
-            {!! Form::textarea('long_description', null, array('class' => 'form-control', 'style' => 'height:200px')) !!}
+            {!! Form::textarea('long_description', (isset($translatable) ? $translatable->long_description : null), array('class' => 'form-control', 'style' => 'height:200px')) !!}
         </div>
     </div>
     @foreach(\Config::get('redminportal::translation') as $translation)
@@ -25,17 +25,29 @@
         <div class="tab-pane" id="lang-{{ $translation['lang'] }}{{ isset($selector_name) ? '-' . $selector_name : '' }}">
             <div class="form-group">
                 {!! Form::label($translation['lang'] . '_name', Lang::get('redminportal::forms.title')) !!}
+                @if (isset($translated))
+                {!! Form::text($translation['lang'] . '_name', (array_key_exists($translation['lang'], $translated) ? $translated[$translation['lang']]->name : ''), array('class' => 'form-control')) !!}
+                @else
                 {!! Form::text($translation['lang'] . '_name', null, array('class' => 'form-control')) !!}
+                @endif
             </div>
-
+            
             <div class="form-group">
                 {!! Form::label($translation['lang'] . '_short_description', Lang::get('redminportal::forms.summary')) !!}
+                @if (isset($translated))
+                {!! Form::text($translation['lang'] . '_short_description', (array_key_exists($translation['lang'], $translated) ? $translated[$translation['lang']]->short_description : ''), array('class' => 'form-control')) !!}
+                @else
                 {!! Form::text($translation['lang'] . '_short_description', null, array('class' => 'form-control')) !!}
+                @endif
             </div>
 
             <div class="form-group">
                 {!! Form::label($translation['lang'] . '_long_description', Lang::get('redminportal::forms.description')) !!}
+                @if (isset($translated))
+                {!! Form::textarea($translation['lang'] . '_long_description', (array_key_exists($translation['lang'], $translated) ? $translated[$translation['lang']]->long_description : ''), array('class' => 'form-control', 'style' => 'height:200px')) !!}
+                @else
                 {!! Form::textarea($translation['lang'] . '_long_description', null, array('class' => 'form-control', 'style' => 'height:200px')) !!}
+                @endif
             </div>
         </div>
         @endif