+++ CHANGELOG.md
@@ -21,6 +21,7 @@ This update introduces some new tables. Refer to [UPGRADE.md](UPGRADE.md) for th
 
 ### New features:
 1. Added product variations. You can now add variations to each product. (issue #125)
+2. Added reusable partial templates for Modal Window and Language selector tab.
 
 ### Enhancements: